requirements.txt
absl-py==2.1.0
ale-py==0.8.1
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
atomicwrites==1.4.1
attrs==23.2.0
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
Bottleneck @ file:///C:/b/abs_f05kqh7yvj/croot/bottleneck_1707864273291/work
box2d-py==2.3.5
Brotli @ file:///C:/b/abs_3d36mno480/croot/brotli-split_1714483178642/work
certifi @ file:///C:/b/abs_1fw_exq1si/croot/certifi_1725551736618/work/certifi
cffi==1.16.0
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
cloudpickle==3.0.0
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1710320294760/work
contourpy @ file:///C:/b/abs_d5rpy288vc/croots/recipe/contourpy_1663827418189/work
cramjam==2.8.3
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
Cython==3.0.10
debugpy @ file:///D:/bld/debugpy_1707444572519/work
decorator==4.4.2
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1698579936712/work
faiss==1.8.0
fasteners==0.19
fastparquet==2024.2.0
filelock @ file:///C:/b/abs_f2gie28u58/croot/filelock_1700591233643/work
fonttools @ file:///C:/b/abs_f47gnfqnx0/croot/fonttools_1713551644747/work
fsspec==2024.6.0
glfw==2.7.0
gmpy2 @ file:///C:/ci/gmpy2_1645456279018/work
graphviz==0.20.3
gym==0.26.2
gym-notices==0.0.8
h5py @ file:///C:/b/abs_c4ha_1xv14/croot/h5py_1715094776210/work
hiddenlayer==0.3
idna @ file:///C:/b/abs_aad84bnnw5/croot/idna_1714398896795/work
imageio==2.34.1
imageio-ffmpeg==0.5.1
importlib-resources @ file:///C:/b/abs_d0dmp77t95/croot/importlib_resources-suite_1704281892795/work
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1710971335535/work
iniconfig==2.0.0
ipykernel @ file:///D:/bld/ipykernel_1708996677248/work
ipython @ file:///D:/bld/ipython_1683289169355/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
Jinja2 @ file:///C:/b/abs_92fccttino/croot/jinja2_1716993447201/work
joblib==1.4.2
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1716472197302/work
jupyter_core @ file:///D:/bld/jupyter_core_1710257325428/work
kagglehub @ file:///C:/b/abs_08f8j6y1m6/croot/kagglehub_1720558573125/work
kiwisolver @ file:///C:/b/abs_88mdhvtahm/croot/kiwisolver_1672387921783/work
lazy_loader==0.4
llvmlite==0.41.1
lmdb==1.4.1
lz4==4.3.3
MarkupSafe @ file:///C:/b/abs_ecfdqh67b_/croot/markupsafe_1704206030535/work
matplotlib @ file:///C:/b/abs_085jhivdha/croot/matplotlib-suite_1693812524572/work
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1713250518406/work
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186184278/work
mkl-service==2.4.0
moviepy==1.0.3
mpmath @ file:///C:/b/abs_7833jrbiox/croot/mpmath_1690848321154/work
mujoco==2.2.0
mujoco-py==2.1.2.14
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
networkx @ file:///C:/b/abs_e6gi1go5op/croot/networkx_1690562046966/work
numba==0.58.1
numexpr @ file:///C:/b/abs_a7kbak88hk/croot/numexpr_1668713882979/work
numpy @ file:///C:/b/abs_54abayvc9j/croot/numpy_and_numpy_base_1682520598361/work
opencv-python==4.9.0.80
packaging @ file:///C:/b/abs_cc1h2xfosn/croot/packaging_1710807447479/work
pandas @ file:///C:/miniconda3/conda-bld/pandas_1692299636855/work
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1712320355065/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
pillow @ file:///C:/b/abs_95ls_qh4c2/croot/pillow_1714398870615/work
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1715777629804/work
pluggy==1.5.0
ply==3.11
pooch @ file:///C:/b/abs_a8nmng7d_x/croot/pooch_1695850149827/work
proglog==0.1.10
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1702399386289/work
psutil @ file:///D:/bld/psutil_1705722577218/work
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
py==1.11.0
pyarrow==16.1.0
pycparser==2.22
pygame==2.1.0
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1714846767233/work
pynndescent==0.5.13
PyOpenGL==3.1.7
pyparsing @ file:///C:/Users/BUILDE~1/AppData/Local/Temp/abs_7f_7lba6rl/croots/recipe/pyparsing_1661452540662/work
PyQt5==5.15.10
PyQt5-sip @ file:///C:/b/abs_c0pi2mimq3/croot/pyqt-split_1698769125270/work/pyqt_sip
PySocks @ file:///C:/ci/pysocks_1605287845585/work
pytest==7.0.1
python-dateutil @ file:///C:/b/abs_3au_koqnbs/croot/python-dateutil_1716495777160/work
pytz @ file:///C:/b/abs_6ap4tsz1ox/croot/pytz_1713974360290/work
PyWavelets==1.4.1
pywin32==306
PyYAML @ file:///C:/b/abs_14xkfs39bx/croot/pyyaml_1728657968772/work
pyzmq @ file:///D:/bld/pyzmq_1715024561746/work
requests @ file:///C:/b/abs_9c5n24p1y7/croot/requests_1716902867639/work
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
sip @ file:///C:/b/abs_edevan3fce/croot/sip_1698675983372/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
swig==4.2.1
sympy @ file:///C:/b/abs_4e4p71hdj_/croot/sympy_1724938208509/work
threadpoolctl==3.5.0
tifffile==2023.7.10
tomli @ file:///C:/Windows/TEMP/abs_ac109f85-a7b3-4b4d-bcfd-52622eceddf0hy332ojo/croots/recipe/tomli_1657175513137/work
torch==2.3.1
torchaudio==2.3.1
torchinfo @ file:///home/conda/feedstock_root/build_artifacts/torchinfo_1684097133203/work
torchview==0.2.7
torchvision==0.18.1
tornado @ file:///C:/b/abs_0cbrstidzg/croot/tornado_1696937003724/work
tqdm @ file:///C:/b/abs_a7h7n45ccq/croot/tqdm_1716395987133/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1713535121073/work
typing_extensions @ file:///C:/b/abs_0as9mdbkfl/croot/typing_extensions_1715268906610/work
tzdata @ file:///croot/python-tzdata_1690578112552/work
umap-learn==0.5.7
unicodedata2 @ file:///C:/b/abs_b6apldlg7y/croot/unicodedata2_1713212998255/work
urllib3 @ file:///C:/b/abs_8e4z8_gh1l/croot/urllib3_1715636317140/work
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1704731205417/work
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306167264/work
zipp @ file:///C:/b/abs_b0beoc27oa/croot/zipp_1704206963359/work
